fix(backend): resolve all build errors — complete missing architecture layers
- Employee entity: added FullName, Username, PasswordHash, StoreId properties
- Added Attendance entity with clock-in/out tracking
- Added DbSet<Attendance> + EF configuration to AppDbContext
- Created 3 domain interfaces: IProductRepository, ICustomerRepository, IBillRepository
- Created 5 application interfaces: IProductService, ICustomerService, IBillService, IInventoryService, IAuthService
- Created 5 application services implementing the interfaces
- Created 3 infrastructure repositories implementing domain interfaces
- Fixed double*decimal type mismatches in BillsController + BillRepository
- Fixed NotImplementedException in AuthController.RefreshToken
- Fixed nullable TimeSpan.TotalHours in EmployeesController

Build result: 0 warnings, 0 errors

+++ SS_MART_API/Controllers/BillsController.cs
@@ -182,9 +182,9 @@ public class BillsController : ControllerBase
             BillNumber = await GenerateBillNumber(),
             CustomerId = originalBill.CustomerId,
             BillDate = DateTime.UtcNow,
-            Subtotal = request.Items.Sum(i => i.Quantity * i.UnitPrice),
+            Subtotal = request.Items.Sum(i => (decimal)i.Quantity * i.UnitPrice),
             TaxAmount = 0,
-            TotalAmount = request.Items.Sum(i => i.Quantity * i.UnitPrice),
+            TotalAmount = request.Items.Sum(i => (decimal)i.Quantity * i.UnitPrice),
             PaymentMode = "CASH",
             Status = "completed",
             IsReturn = true,
@@ -198,7 +198,7 @@ public class BillsController : ControllerBase
                 ProductId = i.ProductId,
                 Quantity = i.Quantity,
                 UnitPrice = i.UnitPrice,
-                TotalAmount = i.Quantity * i.UnitPrice
+                TotalAmount = (decimal)i.Quantity * i.UnitPrice
             }).ToList()
         };
